code/requirements.txt
imbalanced_learn==0.10.1
joblib==1.1.1
matplotlib==3.7.1
networkx==2.8.4
numpy==1.23.5
pandas==1.5.3
python_igraph==0.11.4
PyWavelets==1.4.1
scikit_learn==1.2.2
scipy==1.10.0
seaborn==0.13.2
xlwings==0.29.1
